update memory rule

+++ ChatBeet/Rules/MemoryCellRule.cs
@@ -25,36 +25,53 @@ namespace ChatBeet.Rules
 
         public override async IAsyncEnumerable<IClientMessage> RespondAsync(PrivateMessage incomingMessage)
         {
-            var setRgx = new Regex($"^({Regex.Escape(config.Nick)}, |{Regex.Escape(config.CommandPrefix)})remember (.*)=(.*)", RegexOptions.IgnoreCase);
+            var setRgx = new Regex($"^({Regex.Escape(config.Nick)}, |{Regex.Escape(config.CommandPrefix)})remember (.*?)=(.*)", RegexOptions.IgnoreCase);
             var setMatch = setRgx.Match(incomingMessage.Message);
             if (setMatch.Success)
             {
                 var key = setMatch.Groups[2].Value.Trim().ToLower();
                 var value = setMatch.Groups[3].Value.Trim();
 
-                var existingCell = await ctx.MemoryCells.FindAsync(key);
-                if (existingCell != null)
-                {
-                    ctx.MemoryCells.Remove(existingCell);
-                    await ctx.SaveChangesAsync();
-                }
-
-                ctx.MemoryCells.Add(new MemoryCell
-                {
-                    Author = incomingMessage.From,
-                    Key = key,
-                    Value = value
-                });
-                await ctx.SaveChangesAsync();
-
-                yield return new PrivateMessage(incomingMessage.GetResponseTarget(), "Got it! 👍");
-
-                if (existingCell != null)
+                if (string.IsNullOrEmpty(key))
                 {
                     yield return new PrivateMessage(
-                        incomingMessage.GetResponseTarget(),
-                        $"Previous value was {IrcValues.BOLD}{existingCell.Value}{IrcValues.RESET}, set by {existingCell.Author}."
-                    );
+                            incomingMessage.GetResponseTarget(),
+                            $"{incomingMessage.From}: provide a name to define."
+                        );
+                }
+                else if (string.IsNullOrEmpty(value))
+                {
+                    yield return new PrivateMessage(
+                            incomingMessage.GetResponseTarget(),
+                            $"{incomingMessage.From}: provide a value to set for {IrcValues.BOLD}{key}{IrcValues.RESET}."
+                        );
+                }
+                else
+                {
+                    var existingCell = await ctx.MemoryCells.FindAsync(key);
+                    if (existingCell != null)
+                    {
+                        ctx.MemoryCells.Remove(existingCell);
+                        await ctx.SaveChangesAsync();
+                    }
+
+                    ctx.MemoryCells.Add(new MemoryCell
+                    {
+                        Author = incomingMessage.From,
+                        Key = key,
+                        Value = value
+                    });
+                    await ctx.SaveChangesAsync();
+
+                    yield return new PrivateMessage(incomingMessage.GetResponseTarget(), "Got it! 👍");
+
+                    if (existingCell != null)
+                    {
+                        yield return new PrivateMessage(
+                            incomingMessage.GetResponseTarget(),
+                            $"Previous value was {IrcValues.BOLD}{existingCell.Value}{IrcValues.RESET}, set by {existingCell.Author}."
+                        );
+                    }
                 }
             }
